Refactor: Change auto-sync interval to hours and update UI
This commit refactors the auto-sync interval to be represented in hours instead of minutes across the application.

Key changes:

- **Data Types & Storage:**
    - `PreferenceManager` and `PreferenceManagerImpl`:
        - `saveAutoSyncInterval` now accepts an `Int` representing hours.
        - `loadAutoSyncInterval` now returns an `Int` representing hours, defaulting to 12 if not set.
    - `SettingsEvent.UpdateSyncInterval`: `intervalInHours` is now `Int`.
    - `SyncUiState`: `autoSyncInterval` is now `Int` and defaults to 1.
    - `SettingsUiState`: `syncIntervalInHours` is now `Int` and defaults to 1.
- **Settings Screen:**
    - `SettingsServerConfigurationDialog`:
        - The `currentIntervalHours` parameter is now used.
        - `intervalOptions` are now `listOf(1, 6, 12, 24)` representing hours.
        - The `OutlinedTextField` and `DropdownMenuItem` now display "hours" instead of "mins".
    - `SettingsScreen.kt`: Replaced `Divider` with `HorizontalDivider` for consistency.
- **Sync Module:**
    - `SyncViewModel`:
        - `updateAutoSyncInterval` now accepts an `Int`.
        - When scheduling periodic sync, `intervalHours` (Int) is converted to `Long` for `syncManager.schedulePeriodicSync`.
    - `SyncScreen.kt`: The auto-sync interval display now appends "hours".
- **ViewModel Logic:**
    - `SettingsViewModel.updateSyncInterval` now correctly handles the interval as `Int` (hours).

The default auto-sync interval is now 1 hour. This change provides a more user-friendly and consistent way to manage auto-sync frequency.

+++ core/src/main/java/com/lyecdevelopers/core/data/preference/PreferenceManager.kt
@@ -41,9 +41,9 @@ interface PreferenceManager {
 
     suspend fun loadAutoSyncEnabled(): Boolean
 
-    suspend fun saveAutoSyncInterval(hours: Long)
+    suspend fun saveAutoSyncInterval(hours: Int)
 
-    suspend fun loadAutoSyncInterval(): Long
+    suspend fun loadAutoSyncInterval(): Int
 
     suspend fun loadServerUrl(): String
